RankingStoreを別ファイルに分離/シングルトン(？)にする
