+++ package.json
@@ -23,6 +23,7 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babelify": "^7.3.0",
+    "eventemitter2": "^4.1.2",
     "isomorphic-fetch": "^2.2.1",
     "laravel-elixir": "^6.0.0-15",
     "react": "^15.6.1",
